moved AuthServiceDatabaseException to package exception
